#43 Moved auth code to auth module.
