+++ api/src/api.py
@@ -2,7 +2,7 @@ import flask_cors
 from flask import Flask, jsonify
 from sqlalchemy.exc import OperationalError
 
-from core.models import AccessToken
+from core.auth import authenticate_request
 from service.config import get_mysql_config
 from service.db import create_mysql_engine, shutdown_session
 from service.error import ApiError, error_handler_api, error_handler_db, error_handler_500, error_handler_404
@@ -26,7 +26,7 @@ app.register_error_handler(ApiError, error_handler_api)
 app.register_error_handler(500, error_handler_500)
 app.register_error_handler(404, error_handler_404)
 app.teardown_appcontext(shutdown_session)
-app.before_request(AccessToken.authenticate_request)
+app.before_request(authenticate_request)
 
 engine = create_mysql_engine(**get_mysql_config())
